Stop building embulk-*.gem distributions, and embed gem structure in embulk-core-*.jar
