--- Gemfile
@@ -1,2 +1,0 @@
-source 'https://rubygems.org/'
-gemspec
