+++ embulk-core/src/main/ruby/embulk/buffer.rb
