+++ embulk-core/src/main/ruby/embulk/column.rb
