+++ embulk-core/src/main/ruby/embulk/data_source.rb
