+++ embulk-core/src/main/ruby/embulk/decoder_plugin.rb
