+++ embulk-core/src/main/ruby/embulk/encoder_plugin.rb
