+++ embulk-core/src/main/ruby/embulk/error.rb
