+++ embulk-core/src/main/ruby/embulk/exec.rb
