+++ embulk-core/src/main/ruby/embulk/executor_plugin.rb
