+++ embulk-core/src/main/ruby/embulk/file_input_plugin.rb
