+++ embulk-core/src/main/ruby/embulk/file_output.rb
