+++ embulk-core/src/main/ruby/embulk/file_output_plugin.rb
