+++ embulk-core/src/main/ruby/embulk/filter_plugin.rb
