+++ embulk-core/src/main/ruby/embulk/formatter_plugin.rb
