+++ embulk-core/src/main/ruby/embulk/guess/bzip2.rb
