+++ embulk-core/src/main/ruby/embulk/guess/csv.rb
