+++ embulk-core/src/main/ruby/embulk/guess/csv_all_strings.rb
