+++ embulk-core/src/main/ruby/embulk/guess/gzip.rb
