+++ embulk-core/src/main/ruby/embulk/guess/json.rb
