+++ embulk-core/src/main/ruby/embulk/guess/newline.rb
