+++ embulk-core/src/main/ruby/embulk/guess/schema_guess.rb
