+++ embulk-core/src/main/ruby/embulk/guess_plugin.rb
