+++ embulk-core/src/main/ruby/embulk/input_plugin.rb
