+++ embulk-core/src/main/ruby/embulk/java/bootstrap.rb
