+++ embulk-core/src/main/ruby/embulk/java/imports.rb
