+++ embulk-core/src/main/ruby/embulk/java_plugin.rb
