+++ embulk-core/src/main/ruby/embulk/logger.rb
