+++ embulk-core/src/main/ruby/embulk/output_plugin.rb
