+++ embulk-core/src/main/ruby/embulk/page.rb
