+++ embulk-core/src/main/ruby/embulk/page_builder.rb
