+++ embulk-core/src/main/ruby/embulk/parser_plugin.rb
