+++ embulk-core/src/main/ruby/embulk/plugin.rb
