+++ embulk-core/src/main/ruby/embulk/plugin_registry.rb
