+++ embulk-core/src/main/ruby/embulk/runner.rb
