+++ embulk-core/src/main/ruby/embulk/schema.rb
