+++ embulk-core/src/main/ruby/embulk/version.rb
