+++ embulk-core/src/test/ruby/monkey_strptime/mri/date/test_date_strptime.rb
