+++ embulk-core/src/test/ruby/vanilla/guess/test_csv_all_strings.rb
