+++ embulk-core/src/test/ruby/vanilla/guess/test_schema_guess.rb
