+++ embulk-core/src/test/ruby/vanilla/run-test.rb
@@ -1,13 +1,14 @@
-base_dir = File.expand_path(File.join(File.dirname(__FILE__), "..", ".."))
-lib_dir = File.join(base_dir, "lib")
-test_dir = File.join(base_dir, "test")
+root_dir = File.expand_path(File.join(File.dirname(__FILE__), "..", "..", "..", "..", ".."))
+src_dir = File.join(root_dir, "embulk-core", "src")
+lib_dir = File.join(src_dir, "main", "ruby")
+test_dir = File.join(src_dir, "test", "ruby")
 
 $LOAD_PATH.unshift(lib_dir)
 $LOAD_PATH.unshift(test_dir)
 
 if ARGV.empty?
-  STDERR.puts "No JAR is specified in ARGV. Looking for Embulk core's JAR in: #{base_dir}/classpath"
-  found_jar_files = Dir.glob("#{base_dir}/classpath/embulk-core-*.jar")
+  STDERR.puts "No JAR is specified in ARGV. Looking for Embulk core's JAR in: #{root_dir}/classpath"
+  found_jar_files = Dir.glob("#{root_dir}/classpath/embulk-core-*.jar")
   if found_jar_files.empty?
     raise IOError, "Embulk core's JAR not found."
   end
@@ -29,7 +30,7 @@ Gem::Specification.reset
 
 require "test/unit"
 
-Dir.glob("#{base_dir}/test/vanilla/**/test{_,-}*.rb") do |file|
+Dir.glob("#{test_dir}/vanilla/**/test{_,-}*.rb") do |file|
   require file.sub(/\.rb$/,"")
 end
