+++ embulk-core/src/test/ruby/vanilla/time/test_timestamp_parser.rb
